requirements.txt
anthropic>=0.18.0
pyyaml>=6.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
